Vss2Git

+++ Source/OldCode/Archive/2005 TVA Code Library/Version/2.0/Source/Tva.Core/Tva.Text/Encoding.vb
@@ -35,7 +35,7 @@ Namespace Text
         ''' <returns>specified Unicode character in proper Regular Expression format</returns>
         Public Shared Function EncodeRegexChar(ByVal item As Char) As String
 
-            Return "\u" & Convert.ToInt32(item).ToString("x"c).PadLeft(4, "0"c)
+            Return "\u" & Convert.ToInt16(item).ToString("x"c).PadLeft(4, "0"c)
 
         End Function
